extension/python/requirements.txt
requests==2.31.0
typer[all]==0.9.0
rich==13.7.0
pydantic==2.5.3
python-dotenv==1.0.0
pytest-asyncio==0.21.1
groq==0.30.0 